tests for adding/removing badges
